requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
distro==1.9.0
Flask==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openai==1.39.0
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
Werkzeug==3.0.3
serverless-wsgi>=2.0.2